refactor: RFC 0004 — extract :core-network module (shared DTOs)
First migration step of the :core extraction: create a :core-network
Android library and move the TBA DTO package into it, so :app and :wear
share one source of truth instead of :wear forking its own copies.

- New :core-network module (Kotlin + serialization, no Android deps).
- Move data.remote.dto (18 files) into it; package preserved so :app
  needs zero import changes and no split packages result.
- Reconcile MediaDto: unify app's top-level base64Image field with
  wear's nested details["base64Image"] read into one flat accessor
  (behavior-preserving for both).
- Delete the 4 forked wear DTOs; repoint WearTbaApi + the complication
  worker at the shared package.

Deferred to follow-up PRs (see RFC body): unify the TbaApi interface +
OkHttp/Retrofit/Json provisioning (fixes the 5MB-vs-20MB cache and
battery-constraint divergence), then :core-domain and :core-data.
