+++ core-network/src/main/kotlin/com/thebluealliance/android/data/remote/dto/ApiStatusDto.kt
