+++ core-network/src/main/kotlin/com/thebluealliance/android/data/remote/dto/EventInsightsDto.kt
