+++ settings.gradle.kts
@@ -17,3 +17,4 @@ rootProject.name = "the-blue-alliance-android"
 include(":app")
 include(":wear")
 include(":tv")
+include(":core-network")
