--- wear/src/main/kotlin/com/thebluealliance/android/wear/data/dto/MatchDto.kt
@@ -1,30 +1,0 @@
-package com.thebluealliance.android.wear.data.dto
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JsonObject
-
-@Serializable
-data class MatchDto(
-    val key: String,
-    @SerialName("comp_level") val compLevel: String,
-    @SerialName("match_number") val matchNumber: Int,
-    @SerialName("set_number") val setNumber: Int,
-    val alliances: MatchAlliancesDto? = null,
-    val time: Long? = null,
-    @SerialName("predicted_time") val predictedTime: Long? = null,
-    @SerialName("winning_alliance") val winningAlliance: String? = null,
-    @SerialName("score_breakdown") val scoreBreakdown: JsonObject? = null,
-)
-
-@Serializable
-data class MatchAlliancesDto(
-    val red: MatchAllianceDto? = null,
-    val blue: MatchAllianceDto? = null,
-)
-
-@Serializable
-data class MatchAllianceDto(
-    val score: Int = -1,
-    @SerialName("team_keys") val teamKeys: List<String> = emptyList(),
-)
